Refactor visualizations for improved type safety and clarity
- Updated type annotations in `graph_barons_heralds.py`, `graph_drakes.py`, `graph_first_bloods.py`, `graph_kills.py`, `jungle_clear_analysis.py`, `objective_analysis.py`, and `personal_performance.py` to enhance static type checking.
- Simplified data extraction logic in `graph_drakes.py` and `graph_kills.py` by using TypedDicts for better structure and readability.
- Improved handling of optional values and added type casting where necessary to prevent runtime errors.
- Enhanced plotting functions to ensure consistent data types are used, particularly for averages and win rates.
- Updated tests in `test_graph_drakes.py`, `test_graph_kills.py`, and others to reflect changes in data structures and ensure type safety.
- Removed unnecessary debug print statements and comments for cleaner code.
- Adjusted descriptions in argparse to be more concise and clear.

+++ stats_visualization/viz_types.py
@@ -32,9 +32,9 @@ class EarlyGameData(TypedDict):
     first_blood_deaths: int
     first_blood_assists: int
     first_tower_kills: int
-    early_kills: List[int]
-    early_deaths: List[int]
-    early_cs: List[int]
+    early_kills: List[float]
+    early_deaths: List[float]
+    early_cs: List[float]
     wins: List[bool]
     champions: List[str]
     roles: List[str]
@@ -63,7 +63,7 @@ class EconomyData(TypedDict):
     cs_per_min: List[float]
     gold_per_min: List[float]
     damage_per_gold: List[float]
-    vision_score: List[int]
+    vision_score: List[float]
     game_durations: List[float]
     total_cs: List[int]
     total_gold: List[int]
